+++ index.js
@@ -15,15 +15,11 @@ var compression = require('compression');
 var bodyParser = require('body-parser');
 var cookieParser = require('cookie-parser');
 var session = require('express-session');
-var fileStore = require('session-file-store')(session);
-var redis = require("redis");
-//var redisStore = require('connect-redis')(session);
-//var client = redis.createClient();
 var cors = require('cors');
 var config = require('./config/config.js');
-
-
+var store;
 var app = express();
+
 app.use(compression());
 app.use(cors());
 app.use(cookieParser());
@@ -37,37 +33,49 @@ app.use(bodyParser.urlencoded({
     limit: '50mb'
 }));
 app.set('trust proxy', 1); // trust first proxy
-app.use(session({
-    store: new fileStore({
+
+if (typeof config.redisHost === "string") {
+    var redis = require("redis");
+    var redisStore = require('connect-redis')(session);
+    var client = redis.createClient({
+        host: config.redisHost.split(":")[0],
+        port: config.redisHost.split(":")[1] || 6379,
+        retry_strategy: function (options) {
+            if (options.error && options.error.code === 'ECONNREFUSED') {
+                return new Error('The server refused the connection');
+            }
+            if (options.total_retry_time > 1000 * 60 * 60) {
+                return new Error('Retry time exhausted');
+            }
+            if (options.attempt > 10) {
+                return undefined;
+            }
+            return Math.min(options.attempt * 100, 3000);
+        }
+    });
+    store = new redisStore({
+        client: client,
+        ttl: 260
+    });
+} else {
+    var fileStore = require('session-file-store')(session);
+    store = new fileStore({
         ttl: 86400,
         logFn: function () {
         },
         path: "/tmp/sessions"
-    }),
-    secret: 'keyboard cat',
-    resave: false,
-    saveUninitialized: false,
-    name: "connect.gc2",
-    cookie: {secure: false}
-}));
+    });
+}
 
-/*
 app.use(session({
-    // create new redis store.
-    store: new redisStore({
-        host: '172.18.0.4',
-        port: 6379,
-        client: client,
-        ttl: 260
-    }),
+    store: store,
     secret: 'keyboard cat',
     resave: false,
     saveUninitialized: false,
     name: "connect.gc2",
     cookie: {secure: false}
-
 }));
-*/
+
 app.use('/app/:db/:schema?', express.static(path.join(__dirname, 'public'), {maxage: '60s'}));
 if (config.staticRoutes) {
     for (var key in config.staticRoutes) {
@@ -86,7 +94,7 @@ const port = process.env.PORT ? process.env.PORT : 3000;
 const server = http.createServer(app);
 if (!sticky.listen(server, port)) {
     // Master code
-    server.once('listening', function() {
+    server.once('listening', function () {
         console.log(`server started on port ${port}`);
     });
 } else {