+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [CalVer](https://calver.org/).
 - Meets the requirements for a Progressive Web App (PWA)   
 - Implemented a generic pool for puppeteer processes, so there is always a warm chromium browser ready for print.
 - Node clustering implemented with default count of workers, which is the cpu core count.
-- Sticky sessions so Socket.io works with clustering.
+- Sticky sessions implemented so Socket.io works with clustering.
+- Redis added as storage for sessions. Just set `redisHost` in `config/config.js`. If not set then file storage will be used.
 
 ### Fixed
 - The startup message set by `startUpModal` is ever shown in print.